+++ tests/sample-plugin.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * WP Plugin Info · Eliasis plugin for WordPress plugins
+ * 
+ * @author     Josantonius - hello@josantonius.com
+ * @copyright  Copyright (c) 2017
+ * @license    GPL-2.0+
+ * @link       https://github.com/Josantonius/WP_Plugin-Info.git
+ * @since      1.0.3
+ */
+
+use Eliasis\App\App,
+	Eliasis\Complement\Type\Plugin\Plugin;
+
+/** 
+ * Start application.
+ */
+App::run('/media/subito/Saturno/Jota/Tecnologies/PHP/Frameworks/Eliasis/complements/Plugins/WordPress/', 'wordpress-plugin');
+